create dedicated credits page
