+++ Changelog.md
@@ -5,6 +5,7 @@
 - add Stats for time on days in week/month
 - Introducing the Statistics tab
 - make "simons bot format" an devmode option (hidden by default)
+- move licenses to own credits page
 
 ## 0.4.0
